Gave up on SSR auth redirecting
Because afte two days of struggling, reading through the docs, messages from ppl on discord and even the source code, I still couldn't get the cookie needed for SSR auth to get set and therefor the SSR auth didn't work. I will and I'm saying WILL work on this later, but for now I'll just keep it as it is and focus on other thing.