Did some polishing all over the app and made notifications dragable
I made the notifications dragable, because I want to be able to swipe them away on both mobile and desktop devices. I tried a few libraries for it, like Hammer.js, svelte-gestures and maybe more, but I figured out that for my usecase it's the best to just code it all from scratch. And it isn't even that hard (so far)!